\name{dropDimer}
\alias{dropDimer}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
find the primer identifiers with self dimers
}
\description{
This function will parse the primer checking report and give out the self-dimer primer identifiers.
}
\usage{
dropDimer(report, ndimer)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{report}{
the report file which was generated by the webtool from Thermo Fisher.
}
  \item{ndimer}{
argument to specify thredshold at which self-dimer primer will be filtered out. If you specify 2, all primer with 2 or more than 2 self dimers will be found out.
}
}
\details{
this function will find out the primer identifiers with two or three self-dimer in the barcoded primers. This function accept ndimer given in numeric values like 2 and 3, or in character 'two' or 'three'.
}
\value{
return a vector containing the self-dimer primer identifiers.
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
Sizhong Yang <yanglzu@163.com>
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{

data(report)
selfdimers = dropDimer(report, ndimer = 3)

## remove the selfdimers from the result

data(PrimerLabels)

wcomb = reparser(report)
tcomb = primerCombi(PrimerLabels)

vcombs = slim(tcomb, wcomb)
moreprimers = pickmore(combi = vcombs, n = 120)
dim(moreprimers)

## further remove the primers with selfdimers

toDrop = apply(moreprimers, 1, function(x) any(match(x, selfdimers, nomatch = 0) > 0))
outprimers = moreprimers[!toDrop,]
dim(outprimers)
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ methods }% use one of  RShowDoc("KEYWORDS")
