\name{pickunique}
\alias{pickunique}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
randomly pick up the combination of primer sets without reported primer dimers.
}
\description{
This function will randomly pick up the unique combination of primer sets. Unique means all the forward and reverse barcoded primer will be unique (used only once). Therefore, the resultant primer sets will be in the same amount as the input PrimerLabels file.
}
\usage{
pickunique(combi)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{combi}{
argument to specify the unique primer set which have been filtered by removing primer dimer conbinations.
}
}
\details{
The file to assigned to argument combi will be generated by slim function in this package. That is a random combination of barcoded primer sets within which the primer dimer sets were removed.
}
\value{
The function will return a data frame of two columns. The 1st and 2nd columns show the identifiers of forward primer and reverse primer, respectively. If combinations more than the amount of input primerlabels are expected, please use the function 'pickmore'.
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
Sizhong Yang <yanglzu@163.com>
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{

data(report)
data(PrimerLabels)
wcomb = reparser(report)
tcomb = primerCombi(PrimerLabels)

combs = slim(tcomb, wcomb)
uprimers = pickunique(combi = combs)
dim(uprimers)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ methods }% use one of  RShowDoc("KEYWORDS")
